WIP rhel setup

+++ setup/setup.sh
@@ -110,8 +110,10 @@ detect_os() {
 		echo "This Setup Script only works for Windows Cygwin, Windows Subsystem for Linux and MacOS"
 		exit -1
 	fi
+  [[ -f /etc/redhat-release ]] &&  export isRedhat=true
 	export osreleaseFile=/proc/sys/kernel/osrelease
 	echo "Checking $osreleaseFile (if present)"
+
 	if [[ -f  $osreleaseFile ]]; then
 		osrelease=`cat $osreleaseFile`
 		echo "Linux OS release is $osrelease"
@@ -246,17 +248,21 @@ install_pkgs() {
     brew install the_silver_searcher fortune cowsay upgrade cowsay jenv leiningen nodejs go rlwrap yarn neovim jenv
     brew cask install skim 
  elif [[ $isLinux == true ]]; then
- echo "** NOTE: If RUNNING BEHIND PROXY, export http_proxy/https_proxy"
- sudo -E add-apt-repository -y ppa:neovim-ppa/unstable
- sudo -E add-apt-repository ppa:webupd8team/java
-  # BEGIN: Nodejs install. Taken from https://nodejs.org/en/download/package-manager/#debian-and-ubuntu-based-linux-distributions
-  curl -sL https://deb.nodesource.com/setup_10.x | sudo -E bash -
-  sudo apt-get install -y nodejs
-  # END : Nodejs install
-  sudo -E apt-get --yes install git curl zsh silversearcher-ag netcat-openbsd dh-autoreconf \
-    autoconf pkg-config tmux fortune-mod cowsay zip unzip python3 python3-pip ruby \
-    vim neovim nodejs rar unrar oracle-java8-installer rlwrap yarn
-  sudo apt-get -y autoremove
+   if [[ $isRedhat == true ]]; then
+     install_pkgs_redhat
+   else
+     echo "** NOTE: If RUNNING BEHIND PROXY, export http_proxy/https_proxy"
+     sudo -E add-apt-repository -y ppa:neovim-ppa/unstable
+     sudo -E add-apt-repository ppa:webupd8team/java
+     # BEGIN: Nodejs install. Taken from https://nodejs.org/en/download/package-manager/#debian-and-ubuntu-based-linux-distributions
+     curl -sL https://deb.nodesource.com/setup_10.x | sudo -E bash -
+     sudo apt-get install -y nodejs
+     # END : Nodejs install
+     sudo -E apt-get --yes install git curl zsh silversearcher-ag netcat-openbsd dh-autoreconf \
+       autoconf pkg-config tmux fortune-mod cowsay zip unzip python3 python3-pip ruby \
+       vim neovim nodejs rar unrar oracle-java8-installer rlwrap yarn
+            sudo apt-get -y autoremove
+     fi
  elif [[ $isCygwin == true ]]; then
    if [[ -f /tmp/apt-cyg ]]; then
      rm /tmp/apt-cyg
@@ -294,6 +300,16 @@ install_pkgs() {
  #  curl -sL --proto-redir -all,https https://raw.githubusercontent.com/zplug/installer/master/installer.zsh | zsh 
  # fi
 
+}
+
+install_pkgs_redhat() {
+  echo "- (install_pkgs_redhat) Using Yum to install packages..."
+  export LANG=en_US.UTF-8
+  export LANGUAGE=en_US.UTF-8
+  export LC_COLLATE=C
+  export LC_CTYPE=en_US.UTF-8
+  sudo yum install -y https://dl.fedoraproject.org/pub/epel/epel-release-latest-7.noarch.rpm
+  sudo yum install -y neovim python3-neovim
 }
 
 setup_go_linux() {
